Starting restructuring code into multiple services Moved Account Service into the new structure
