+++ libs/transaction.js
