+++ services/account/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "virtwallet-account",
+  "version": "1.0.0",
+  "description": "Virtwallet Account Service",
+  "repository": "https://github.com/silvamb/virtwallet-serverless.git",
+  "author": "Matheus Silva <crush.comp04@gmail.com>",
+  "license": "GPLv3",
+  "dependencies": {
+    "libs": "file:../../libs"
+  },
+  "devDependencies": {
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "mocha": "^7.0.0",
+    "serverless-iam-roles-per-function": "^2.0.2",
+    "serverless-mocha-plugin": "^1.11.0"
+  },
+  "scripts": {
+    "test": "mocha ./test/*-test.js"
+  }
+}
