+++ services/account/serverless.yml
@@ -1,0 +1,30 @@
+service: virtwallet-account
+
+provider:
+  name: aws
+  runtime: nodejs12.x
+  stage: ${opt:stage, 'dev'}
+  region: ${opt:stage, 'eu-west-1'}
+  apiGateway:
+    restApiId:
+      'Fn::ImportValue': Virtwallet-${self:provider.stage}-RestApiId
+    restApiRootResourceId:
+      'Fn::ImportValue': Virtwallet-${self:provider.stage}-RootResourceId
+  stackTags:
+    Project: virtwallet
+    Service: account
+
+custom:
+  restApiAuthorizerId: '${cf:virtwallet-rest-api-${self:provider.stage}.VirtwalletRestApiAuthorizerId}'
+
+package:
+  excludeDevDependencies: true
+  exclude:
+    - .git/**
+    - test/**
+    - '*.yml'
+    - '*.json'
+
+functions: ${file(./sls-functions.yml)}
+plugins:
+  - serverless-iam-roles-per-function
