+++ services/account/sls-functions.yml
@@ -1,0 +1,62 @@
+apigw-event-handler:
+  handler: src/accountLambda.handle
+  events:
+    - http:
+        path: account
+        method: post
+        authorizer:
+          type: COGNITO_USER_POOLS
+          authorizerId: ${self:custom.restApiAuthorizerId}
+          scopes:
+            - email
+    - http:
+        path: account
+        method: get
+        authorizer:
+          type: COGNITO_USER_POOLS
+          authorizerId: ${self:custom.restApiAuthorizerId}
+          scopes:
+            - email
+    - http:
+        path: account/{accountId}
+        method: get
+        authorizer:
+          type: COGNITO_USER_POOLS
+          authorizerId: ${self:custom.restApiAuthorizerId}
+          scopes:
+            - email
+        parameters:
+          paths:
+            accountId: true
+    - http:
+        path: account/{accountId}
+        method: put
+        authorizer:
+          type: COGNITO_USER_POOLS
+          authorizerId: ${self:custom.restApiAuthorizerId}
+          scopes:
+            - email
+        parameters:
+          paths:
+            accountId: true
+    - http:
+        path: account/{accountId}
+        method: delete
+        authorizer:
+          type: COGNITO_USER_POOLS
+          authorizerId: ${self:custom.restApiAuthorizerId}
+          scopes:
+            - email
+        parameters:
+          paths:
+            accountId: true
+  iamRoleStatements:
+    - Effect: Allow
+      Action:
+        - dynamodb:Query
+        - dynamodb:Scan
+        - dynamodb:GetItem
+        - dynamodb:PutItem
+        - dynamodb:UpdateItem
+        - dynamodb:DeleteItem
+      Resource: "arn:aws:dynamodb:${opt:region, self:provider.region}:*:table/virtwallet"
